% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{addons_deparsers}
\alias{addons_deparsers}
\title{A list of deparsers to support implicit variable declaration and explicit 'return'}
\usage{
addons_deparsers()
}
\description{
A list of deparsers to support implicit variable declaration and explicit 'return'
}
\note{
This is used as input to \link{compile_r} and \link{compile_exprs}.
}
\examples{
addons_deparsers()

}
