% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-annotate_expr.R
\name{annotate_exprs}
\alias{annotate_exprs}
\title{Parse and annotate expressions}
\usage{
annotate_exprs(x)
}
\arguments{
\item{x}{A character string; the input to parse.}
}
\description{
Parse and annotate expressions
}
