% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{basic_deparsers}
\alias{basic_deparsers}
\title{A minimal list of deparsers for deparsing JavaScript}
\usage{
basic_deparsers()
}
\description{
A minimal list of deparsers for deparsing JavaScript
}
\note{
This is used as input to \link{deparse_js}, \link{compile_r} and \link{compile_exprs}.
}
\examples{
basic_deparsers()

}
