% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-bundle.R
\name{bundle}
\alias{bundle}
\title{Bundle a list of files into a single JavaScript file}
\usage{
bundle(fs)
}
\arguments{
\item{fs}{A character vector; a list of R or JavaScript files.
The R files will be transpiled to JavaScript before bundling.}
}
\description{
Bundle a list of files into a single JavaScript file
}
\examples{
library(sketch)
js <- bundle(c(src("dom"), src("io")))
cat(paste(readLines(js), collapse = "\n"))

}
