% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_optional_args.R
\name{capture_args}
\alias{capture_args}
\title{Split a list of arguments into two sets}
\usage{
capture_args(args, capture)
}
\arguments{
\item{args}{A named list; the list of arguments given by \code{list(...)}.}

\item{capture}{A character vector; the variables to keep.}
}
\description{
Split a list of arguments into two sets
}
