% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-compile.R
\name{compile_active}
\alias{compile_active}
\title{Compile active file in 'RStudio'}
\usage{
compile_active(...)
}
\arguments{
\item{...}{Optional arguments to pass to \code{compile_r}.}
}
\description{
Compile active file in 'RStudio'
}
\examples{
\dontrun{
# At 'RStudio', opens a 'sketch' R file in the editor, then
# run the following:
compile_active()
}

}
