% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_json.R
\name{compile_data}
\alias{compile_data}
\title{Compile a data file into a JS file}
\usage{
compile_data(input, output)
}
\arguments{
\item{input}{A character string; the path to the input file.}

\item{output}{A character string; the path to the output file.}
}
\description{
Compile a data file into a JS file
}
