% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-compile.R
\name{compile_exprs}
\alias{compile_exprs}
\title{Compile R code into JavaScript code}
\usage{
compile_exprs(x, rules = default_rules(), deparsers = default_deparsers())
}
\arguments{
\item{x}{A character string; the expression to transpile to JS.}

\item{rules}{A list of rewriting rules. See [make_rule] for more detail.}

\item{deparsers}{A list of deparsers. See [make_deparser] for more detail.}
}
\value{
A character string.
}
\description{
Compile R code into JavaScript code
}
\examples{
compile_exprs("R + Cpp", list(make_rule('Cpp', 'JS')))
compile_exprs("math.add(a, b)", list(make_rule('math.add', '+')))

}
