% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_p5.R
\name{compile_p5_r}
\alias{compile_p5_r}
\title{Compile p5 R file into a JS file}
\usage{
compile_p5_r(input, output = "")
}
\arguments{
\item{input}{A character string; the input file.}

\item{output}{A character string; the output file. When the
output is "", the result is printed to the standard output.}
}
\description{
Compile p5 R file into a JS file
}
