% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_r}
\alias{compile_r}
\title{Compile a 'sketch' R file into a JS file}
\usage{
compile_r(
  input,
  output = "",
  rules = default_rules(),
  deparsers = default_deparsers()
)
}
\arguments{
\item{input}{A character string; the input file.}

\item{output}{A character string; the output file. When the
output is "", the result is printed to the standard output.}

\item{rules}{A list of rewriting rules. See [make_rule] for more detail.}

\item{deparsers}{A list of deparsers. See [make_deparser] for more detail.}
}
\description{
Compile a 'sketch' R file into a JS file
}
