% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_assets.R
\name{convert_src}
\alias{convert_src}
\title{Load JavaScript / CSS / R Sketch Script / CSV file}
\usage{
convert_src(x)
}
\arguments{
\item{x}{A character string; the link / path to the JS / CSS / R script / CSV file.}

\item{renderer}{function; the function to render the shiny tag object.
Use `htmltools::doRenderTags` for RMD, and `identity` for `html_template`.}
}
\description{
Load JavaScript / CSS / R Sketch Script / CSV file
}
\keyword{internal}
