% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{default_2_deparsers}
\alias{default_2_deparsers}
\title{A list of deparsers with additional features}
\usage{
default_2_deparsers()
}
\description{
Support automatic variable declaration, automatic `return`
and shorthand notation for the DOM module.
}
\note{
lifecycle: experimental

This is used as input to \link{compile_r} and \link{compile_exprs}.
}
\examples{
default_2_deparsers()

}
