% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{default_deparsers}
\alias{default_deparsers}
\title{A list of default deparsers for deparsing JavaScript}
\usage{
default_deparsers()
}
\description{
A list of default deparsers for deparsing JavaScript
}
