% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{default_processors}
\alias{default_processors}
\title{A list of handlers for processing the '#!'/'#|' header}
\usage{
default_processors()
}
\description{
A list of handlers for processing the '#!'/'#|' header
}
\note{
This is used as input to \link{assets}.
}
\examples{
default_processors()
}
