% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{deparse_df_arg}
\alias{deparse_df_arg}
\title{Deparser for the arguments in a "data.frame" call}
\usage{
deparse_df_arg(list0, err_msg, ...)
}
\description{
Deparser for the arguments in a "data.frame" call
}
\keyword{internal}
