% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse.R
\name{deparse_js}
\alias{deparse_js}
\title{Expression Deparsing for JS}
\usage{
deparse_js(ast, deparsers)
}
\arguments{
\item{ast}{language object.}

\item{deparsers}{A list of "typed" deparsers.}
}
\value{
A character string.
}
\description{
Expression Deparsing for JS
}
