% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-source_map.R
\name{deparse_js_ast}
\alias{deparse_js_ast}
\title{Deparse a compiled AST}
\usage{
deparse_js_ast(ast)
}
\arguments{
\item{ast}{The compiled AST. The JavaScript AST compiled from the R AST.}
}
\value{
A character string. The compiled string.
}
\description{
Deparse a compiled AST
}
\note{
This feature is experimental.
}
