% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{deparse_list_arg}
\alias{deparse_list_arg}
\title{Deparser for the arguments in a "list" call}
\usage{
deparse_list_arg(list0, err_msg, ...)
}
\description{
Deparser for the arguments in a "list" call
}
\keyword{internal}
