% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_df.R
\name{desc_df}
\alias{desc_df}
\title{Describe a data frame}
\usage{
desc_df(x, max_p_na = 0.99, min_distinct_num = 10,
  max_distinct_cat = 30, min_distinct_mode = 100, max_p_mode = 0.99,
  verbose = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{max_p_na}{???}

\item{min_distinct_num}{integer - a warning will be printed if any \code{dbl} or \code{int}
variables have fewer than this number of distinct (non-negative) values}

\item{max_distinct_cat}{integer - a warning will be printed if any \code{chr}, \code{fct}
or \code{ord} variables have fewer than this number of distinct (non-negative) values}

\item{min_distinct_mode}{integer - mode statistics will not be calculate for variables
with more than this number of distinct (non-negative) values}

\item{max_p_mode}{proportion - a warning will be printed if the proportion of the model
value for any variables exceeds this value}

\item{verbose}{if \code{TRUE} messages will be printed to the console}
}
\description{
Returns a \link[dplyr]{tibble} with rows corresponding to the variables in
\code{x} and columns giving values for a range of summary statistics. If
\code{verbose = TRUE} comments are printed to the console to highlight possible
issues with data that might warrent further inspection prior to analysis.
}
\examples{
\dontrun{
desc_df(iris)
desc_df(mtcars)
desc_df(esoph)
desc_df(attenu)
desc_df(nycflights13::airlines)
desc_df(nycflights13::airports)
desc_df(nycflights13::flights)
desc_df(nycflights13::planes)
desc_df(nycflights13::weather)
}
}
