% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_optional_args.R
\name{do_call}
\alias{do_call}
\title{Execute a function call with extended arguments}
\usage{
do_call(what, ..., extended_args)
}
\arguments{
\item{what}{A function.}

\item{...}{Arguments to the function.}

\item{extended_args}{A list of arguments to the function call.}
}
\description{
Execute a function call with extended arguments
}
\keyword{internal}
