% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{dp}
\alias{dp}
\title{Constructor function to combine low-level deparsers}
\usage{
dp(...)
}
\arguments{
\item{...}{character strings indicating the features needed
of the deparsers. The supported features are "basic", "r",
"auto" and "dom" corresponding to the basic deparsers, the R
support, the automatic variable declaration and return, and
the dom shorthand notation.}
}
\description{
Constructor function to combine low-level deparsers
}
\note{
lifecycle: experimental
}
