% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{load_library}
\alias{load_library}
\alias{load_script}
\title{Header functions}
\usage{
load_library(package, ...)

load_script(src, ...)
}
\arguments{
\item{package}{A character string; name of a JavaScript library.}

\item{...}{Additional arguments to pass to header processor.}

\item{src}{A character string; the full web/local path to a JavaScript library.}
}
\description{
Header functions
}
