% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-rmarkdown.R
\name{eng_sketch}
\alias{eng_sketch}
\title{Knitr engine for 'sketch'}
\usage{
eng_sketch(options)
}
\arguments{
\item{options}{A list of chunk options.}
}
\description{
Knitr engine for 'sketch'
}
\examples{
\dontrun{
# In the 'setup' code chunk of an R Markdown document
knitr::knit_engines$set(sketch = sketch::eng_sketch)
}

}
