% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-bundle.R
\name{flatten_filelist}
\alias{flatten_filelist}
\title{Flatten a list of files and directories into a list of files}
\usage{
flatten_filelist(fs, pattern = NULL, ...)
}
\arguments{
\item{fs}{A character vector; a list of files.}

\item{pattern}{An optional regular expression to pass to `list.files` for
filtering files while expanding a directory into a list of files.}

\item{...}{Additional parameters to pass to `list.files`.}
}
\description{
Flatten a list of files and directories into a list of files
}
\examples{
modules_dir <- system.file("modules", package = "sketch")
flatten_filelist(modules_dir)

}
