% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freduce}
\alias{freduce}
\title{Apply a list of functions sequentially}
\usage{
freduce(value, function_list, preprocess = identity,
  postprocess = identity)
}
\arguments{
\item{value}{initial value.}

\item{function_list}{a list of functions.}

\item{preprocess}{a function for preprocessing the value.}

\item{postprocess}{a function for postprocessing the value.}
}
\description{
Apply a list of functions sequentially
}
\keyword{internal}
