% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-test.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Extract the content of the `load_script` headers of a sketch R file}
\usage{
get_dependencies(app_script, local_only = TRUE)
}
\arguments{
\item{app_script}{A character string; the path to the sketch R file.}

\item{local_only}{TRUE / FALSE; if TRUE, exclude the ones that are web link.}
}
\description{
Extract the content of the `load_script` headers of a sketch R file
}
