% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{load_Family}
\alias{load_Family}
\alias{load_library}
\alias{load_script}
\alias{load_data}
\title{Header functions}
\usage{
load_library(package, ...)

load_script(src, ...)

load_data(x, cache = tempfile(), ...)
}
\arguments{
\item{package}{A character string; name of a JavaScript library.}

\item{...}{Additional arguments to pass to header processor.}

\item{src}{A character string; the full web/local path to a JavaScript library.}

\item{x}{A character string; the full path to the file containing the data.}

\item{cache}{A character string; the full path to the cache file.}
}
\description{
Header functions
}
