% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-source_file.R
\name{html_tags}
\alias{html_tags}
\alias{default_tags}
\alias{basic_tags}
\title{HTML templates}
\usage{
default_tags()

basic_tags()
}
\description{
A list of shiny tags describing a HTML template. The
list must have signature / structure of a named list:
    \code{[head = [shiny.tag], body = [shiny.tag]]}
}
\examples{
default_tags()

basic_tags()
}
