% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-components.R
\name{is_call}
\alias{is_call}
\title{Predicate for calls}
\usage{
is_call(x, name = NULL, n = NULL, ns = NULL)
}
\arguments{
\item{x}{An object to test. If a formula, the right-hand side is
extracted.}

\item{name}{An optional name that the call should match. It is
passed to \code{\link[rlang:sym]{sym()}} before matching. This argument is vectorised
and you can supply a vector of names to match. In this case,
\code{is_call()} returns \code{TRUE} if at least one name matches.}

\item{n}{An optional number of arguments that the call should
match.}

\item{ns}{The namespace of the call. If \code{NULL}, the namespace
doesn't participate in the pattern-matching. If an empty string
\code{""} and \code{x} is a namespaced call, \code{is_call()} returns
\code{FALSE}. If any other string, \code{is_call()} checks that \code{x} is
namespaced within \code{ns}.

Can be a character vector of namespaces, in which case the call
has to match at least one of them, otherwise \code{is_call()} returns
\code{FALSE}.}
}
\description{
Predicate for calls
}
