% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-components.R
\name{is_syntactic_literal}
\alias{is_syntactic_literal}
\title{Predicate for syntactic literal}
\usage{
is_syntactic_literal(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Predicate for syntactic literal
}
\note{
This function is imported from `rlang`.
}
