% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{license_info}
\alias{license_info}
\title{License information}
\usage{
license_info(x)
}
\arguments{
\item{x}{A character string; name of the library / assets.}
}
\value{
A named list containing the license information and
the link from which the information is extracted.
}
\description{
License information
}
\examples{
license_info("mathjs")
license_info("p5")

}
