% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_sketch_in_RMD.R
\name{load_js}
\alias{load_js}
\title{Load js library into an RMD document}
\usage{
load_js(src)
}
\arguments{
\item{src}{A character string; the link to the js library.}
}
\description{
Load js library into an RMD document
}
\examples{
\dontrun{
load_js(src = src("p5"))
load_js(src = "https://cdn.plot.ly/plotly-latest.min.js")
}
}
