% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_sketch_in_RMD.R
\name{load_script}
\alias{load_script}
\title{Load JavaScript / CSS / R Sketch Script / CSV file into an RMD document}
\usage{
load_script(x)
}
\arguments{
\item{x}{A character string; the link / path to the JS / CSS / R script / CSV file}
}
\description{
Load JavaScript / CSS / R Sketch Script / CSV file into an RMD document
}
\examples{
\dontrun{
load_script(src("p5"))
load_script("https://cdn.plot.ly/plotly-latest.min.js")
}
}
