% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{make_processor}
\alias{make_processor}
\title{Make a handle to process header}
\usage{
make_processor(pred, fun)
}
\arguments{
\item{pred}{A function, taking a string and returning a logical.}

\item{fun}{A function, taking a string and returning a 'shiny.tag' object.}
}
\value{
A header processor / handler.
}
\description{
Make a handle to process header
}
