% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_expr}
\alias{parse_expr}
\title{Parse R code}
\usage{
parse_expr(x)
}
\arguments{
\item{x}{Text containing expressions to parse_expr for
\code{parse_expr()} and \code{parse_exprs()}. Can also be an R connection,
for instance to a file. If the supplied connection is not open,
it will be automatically closed and destroyed.}
}
\description{
Parse R code
}
\note{
This function is imported from `rlang`.
}
