% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot_var}
\alias{plot_var}
\title{Plot a variable}
\usage{
plot_var(data, x, group, min_distinct_xnum = 10,
  max_distinct_xcat = 50, max_distinct_group = 3, bins = 30)
}
\arguments{
\item{data}{a data frame}

\item{x}{variable for which the distribution will be plotted}

\item{group}{optional grouping variable}

\item{min_distinct_xnum}{integer, if \code{x} is non-categorical and the number of
distinct values in \code{x} is less than \code{min_distinct_xnum} a error will be
thrown (see Details)}

\item{max_distinct_xcat}{integer, if \code{x} is categorical and the number of distinct
values in \code{x} exceeds \code{max_distinct_xcat} a error will be thrown (see
Details)}

\item{max_distinct_group}{integer, if the number of distinct values in \code{group}
exceeds \code{max_distinct_group} a error will be thrown}

\item{bins}{passed to \link[ggplot2]{geom_histogram} (ignored if \code{x} is
categorical)}
}
\description{
Plot the distribution of a variable and (optionally) facet the plot by a
grouping variable.

The \code{x} and \code{group} arguments can be unquoted column names in \code{data},
or expressions representing functions of one or more column names in \code{data}.
These expressions are passed to \link[dplyr]{transmute} which is used to generate the
data for plotting.
}
\details{
\link[dplyr]{type_sum} is used to determine the class of the \code{x}
variable:
\itemize{
\item if \code{x} is one of \code{"chr"}, \code{"fct"}, \code{"lgl"} or
\code{"ord"} then it is treated as categorical
\item if \code{x} is one of \code{"dbl"} or \code{"int"} then it is treated as
non-categorical
\item no other variable types are supported for \code{x}
}

\link[ggplot2]{geom_histogram} is used to plot non-categorical x variables and
\link[ggplot2]{geom_bar} is used to plot categorical x variables.
}
\examples{
\dontrun{

mtcars \%>\% plot_var(cyl)
mtcars \%>\% plot_var(cyl / disp)
mtcars \%>\% plot_var(cyl / disp, vs + am)
}
}
