% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.R
\name{print.rule}
\alias{print.rule}
\title{Print function for 'rule' objects}
\usage{
\method{print}{rule}(x, ...)
}
\arguments{
\item{x}{A 'rule' object.}

\item{...}{(Unused) Optional arguments.}
}
\description{
Print function for 'rule' objects
}
