% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rewrite.R
\name{print.sketch_rule}
\alias{print.sketch_rule}
\title{Print function for 'sketch_rule' objects}
\usage{
\method{print}{sketch_rule}(x, ...)
}
\arguments{
\item{x}{A 'sketch_rule' object.}

\item{...}{(Unused) Optional arguments.}
}
\description{
Print function for 'sketch_rule' objects
}
\examples{
library(sketch)
rule_1 <- make_rule("+", "Math.add")
print(rule_1)

}
