% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_p5.R
\name{rewrite_ast}
\alias{rewrite_ast}
\title{Rewrite an AST}
\usage{
rewrite_ast(ast, rules)
}
\arguments{
\item{ast}{A language object}

\item{rules}{A list of functions}
}
\description{
Rewrite an AST
}
\keyword{internal}
