% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_p5.R
\name{rewrite_str}
\alias{rewrite_str}
\title{Rewrite a string}
\usage{
rewrite_str(str, rules)
}
\arguments{
\item{str}{A character string}

\item{rules}{A list of functions}
}
\description{
Rewrite a string
}
\keyword{internal}
