% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch-package.R
\docType{package}
\name{sketch-package}
\alias{sketch}
\alias{sketch-package}
\title{sketch: quick and convenient descriptions of data}
\description{
Tools for summarising and visualising variables in
    dataframes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkidney/sketch}
  \item Report bugs at \url{https://github.com/dkidney/sketch/issues}
}

}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@yahoo.co.uk}

}
