% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-source_file.R
\name{source_active}
\alias{source_active}
\title{Source active file in RStudio}
\usage{
source_active(launch_browser = "viewer")
}
\arguments{
\item{launch_browser}{A character string; "viewer" or "browser", which
calls `rstudioapi::viewer` and `utils::browserURL` respectively; use
`NULL` to suppress display.}
}
\description{
Source active file in RStudio
}
