% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_file.R
\name{source_js}
\alias{source_js}
\title{Source a p5.js file}
\usage{
source_js(file, ...)
}
\arguments{
\item{file}{A character string; path to the JS file.}

\item{...}{An optional list of shiny tags to be added to the <head> of
the html template.}
}
\description{
Source a p5.js file
}
