% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-source_map.R
\name{source_map}
\alias{source_map}
\title{Convert a compiled AST into a source map}
\usage{
source_map(ast)
}
\arguments{
\item{ast}{The compiled AST. The JavaScript AST compiled from the R AST.}
}
\value{
A (list of) source map.
}
\description{
Convert a compiled AST into a source map
}
\note{
This feature is experimental.
}
