% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-source_map.R
\name{source_map_from_files}
\alias{source_map_from_files}
\title{Create a source map (.map) file}
\usage{
source_map_from_files(source_file, target_file, ...)
}
\description{
Create a source map (.map) file
}
