% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-source_map.R
\name{source_map_table}
\alias{source_map_table}
\title{Display the source map in a table}
\usage{
source_map_table(x)
}
\arguments{
\item{x}{A source map.}
}
\value{
A data frame.
}
\description{
Display the source map in a table
}
