% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_p5.R
\name{source_p5_js}
\alias{source_p5_js}
\title{Source a p5.js file}
\usage{
source_p5_js(index_js)
}
\arguments{
\item{index_js}{A character string; path to the JS file.}
}
\description{
Source a p5.js file
}
