% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_p5.R
\name{source_p5_r}
\alias{source_p5_r}
\title{Source a p5.R file}
\usage{
source_p5_r(index_r, debug = F)
}
\arguments{
\item{index_r}{A character string; path to the R file.}

\item{debug}{T or F; if T, print compiled code on screen.}
}
\description{
Source a p5.R file
}
