% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-source_file.R
\name{source_r}
\alias{source_r}
\title{Source a sketch R file}
\usage{
source_r(file, debug = F, launch_browser = "viewer")
}
\arguments{
\item{file}{A character string; path to the R file.}

\item{debug}{T or F; if T, print compiled code on screen.}

\item{launch_browser}{A character string; "viewer" or "browser", which
calls `rstudioapi::viewer` and `utils::browserURL` respectively; use
"NULL" (character string) to suppress display.}
}
\description{
This function compiles a sketch R file, resolves the
dependencies and serves it in the viewer.
}
