% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{src}
\alias{src}
\title{Get the source link of a JavaScript library}
\usage{
src(x)
}
\arguments{
\item{x}{A character string; name of the JavaScript library}
}
\value{
A character string; the path to the library.
}
\description{
Get the source link of a JavaScript library
}
\examples{
src("mathjs")
src("p5")

}
