% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{subst}
\alias{subst}
\title{Search a list for 'pattern' and replace each occurrence by 'replacement'}
\usage{
subst(x, pattern, replacement)
}
\arguments{
\item{x}{A list. The list to search.}

\item{pattern}{A character string. The pattern to search for.}

\item{replacement}{A character string. The replacement.}
}
\description{
Search a list for 'pattern' and replace each occurrence by 'replacement'
}
