% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_json.R
\name{to_json}
\alias{to_json}
\title{Convert a file into JS expression}
\usage{
to_json(input, as_data_frame, read_fun, ...)
}
\arguments{
\item{input}{A character string; the path to the input file.}

\item{as_data_frame}{TRUE or FALSE; whether the data are loaded as a data-frame.}

\item{read_fun}{A function to load the input file. Default settings are provided for
CSV files and JSON files. The function has to load a data file into an object that can
be handled by `jsonlite::toJSON`. Possible choices include `utils::read_delim`,
`readr::read_csv2`, etc.}

\item{...}{Extra arguments to be passed to `read_fun`.}
}
\description{
It supports csv and json by default and lets users provide
custom handlers if other file formats are used.
}
