% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{to_shiny_tag}
\alias{to_shiny_tag}
\title{Load JavaScript / CSS / R Sketch Script / CSV file}
\usage{
to_shiny_tag(src, ...)
}
\arguments{
\item{x}{A character string; the link / path to the JS / CSS / R script / CSV file.}
}
\description{
Load JavaScript / CSS / R Sketch Script / CSV file
}
\keyword{internal}
