% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_json.R
\name{compile_data}
\alias{compile_data}
\title{Compile a data file into a JavaScript file}
\usage{
compile_data(input, output = tempfile(), ...)
}
\arguments{
\item{input}{A character string; the path to the input file.}

\item{output}{A character string; the path to the output file.}

\item{...}{Extra arguments to be passed to `to_json`.}
}
\description{
Compile a data file into a JavaScript file
}
\examples{
\dontrun{
file <- system.file("test_files/test_csv.csv", package = "sketch")
readLines(compile_data(file))
}

}
