% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_file.R
\name{compile_r}
\alias{compile_r}
\title{Compile an R file into a JS file}
\usage{
compile_r(input, output = "")
}
\arguments{
\item{input}{A character string; the input file.}

\item{output}{A character string; the output file. When the
output is "", the result is printed to the standard output.}
}
\description{
Compile an R file into a JS file
}
