% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{convert_src}
\alias{convert_src}
\title{Convert an asset link into a 'shiny.tag' object}
\usage{
convert_src(x, processors = default_processors())
}
\arguments{
\item{x}{A character string; the header line (without the prefix #!).}

\item{processors}{A list of handlers for processing the '#!' header.}
}
\value{
A shiny.tag object.
}
\description{
Convert an asset link into a 'shiny.tag' object
}
