% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{default_2_deparsers}
\alias{default_2_deparsers}
\title{A list of deparsers to support implicit variable declaration and explicit 'return'}
\usage{
default_2_deparsers()
}
\description{
\lifecycle{experimental}
}
\note{
This is used as input to \link{compile_r} and \link{compile_exprs}.
}
\examples{
default_2_deparsers()

}
