% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{default_deparsers}
\alias{default_deparsers}
\title{A list of default deparsers for deparsing JavaScript}
\usage{
default_deparsers()
}
\description{
A list of default deparsers for deparsing JavaScript
}
\note{
This is used as input to \link{compile_r} and \link{compile_exprs}.
}
\examples{
\dontrun{
default_deparsers()
}

}
