% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets-to_shiny_tag.R
\name{default_processors}
\alias{default_processors}
\title{List of handlers for processing the '#!' header}
\usage{
default_processors()
}
\description{
List of handlers for processing the '#!' header
}
