% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-components.R
\name{deparse_sym}
\alias{deparse_sym}
\alias{deparse_call}
\alias{deparse_infix}
\alias{deparse_wrap}
\alias{deparse_for}
\alias{deparse_if}
\alias{deparse_while}
\alias{deparse_function}
\alias{deparse_list}
\alias{deparse_df}
\alias{deparse_new}
\alias{deparse_let}
\alias{deparse_dataURI}
\alias{deparse_ifelse}
\alias{deparse_lambda}
\alias{deparse_add}
\alias{deparse_subtract}
\alias{deparse_extract}
\alias{deparse_extractAssign}
\alias{deparse_extract2}
\alias{deparse_extract2Assign}
\title{Deparsers (specialised)}
\usage{
deparse_sym(ast, ...)

deparse_call(ast, ...)

deparse_infix(ast, ...)

deparse_wrap(ast, ...)

deparse_for(ast, ...)

deparse_if(ast, ...)

deparse_while(ast, ...)

deparse_function(ast, ...)

deparse_list(ast, ...)

deparse_df(ast, ...)

deparse_new(ast, ...)

deparse_let(ast, ...)

deparse_dataURI(ast, ...)

deparse_ifelse(ast, ...)

deparse_lambda(ast, ...)

deparse_add(ast, ...)

deparse_subtract(ast, ...)

deparse_extract(ast, ...)

deparse_extractAssign(ast, ...)

deparse_extract2(ast, ...)

deparse_extract2Assign(ast, ...)
}
\arguments{
\item{ast}{A language object.}

\item{...}{The contextual information to be passed on to
the next call.}
}
\value{
A character string.
}
\description{
Deparsers (specialised)

Deparser for calls

Deparser for infix operators

Deparser for brackets

Deparser for the 'for' keyword

Deparser for the 'if' keyword

Deparser for the 'while' keyword

Deparser for the 'function' keyword

Deparser for the "list" operators

Deparser for the "data.frame" operators

Deparser for the "new" operator

Deparser for the "let" operator

Deparser for the "dataURI" operator

Deparser for the "ifelse" operator

Deparser for the "lambda" operator

Deparser for the "add" operator

Deparser for the "subtract" operator

Deparser for the "extract" operator

Deparser for the "extractAssign" operator

Deparser for the "extract2" operator

Deparser for the "extract2Assign" operator
}
