% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{let-declare-const}
\alias{let-declare-const}
\alias{let}
\alias{declare}
\alias{const}
\title{Empty functions}
\usage{
let(...)

declare(...)

const(...)
}
\arguments{
\item{...}{Any arguments}
}
\description{
These functions do nothing. It is created to ensure
the keywords `let` and `declare` are defined.
}
\examples{
let (x)
let (x = 1, y = 2)
declare (x1, x2, x3)

}
