% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{let}
\alias{let}
\title{Empty functions}
\usage{
let(...)
}
\arguments{
\item{...}{Any arguments}
}
\description{
This function does nothing. It is created to ensure
the keywords `let` and `declare` are defined.
}
