% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-rmarkdown.R
\name{eng_sketch}
\alias{eng_sketch}
\title{A language engine for 'sketch'}
\usage{
eng_sketch(options)
}
\arguments{
\item{options}{A list of chunk options.}
}
\description{
This supports the use of 'sketch' code chunk in an R Markdown document.
}
\examples{
# The following line makes `sketch::eng_sketch` available to `knitr::knit_engines`.
# It is usually used in the 'setup' code chunk of an R Markdown document
knitr::knit_engines$set(sketch = sketch::eng_sketch)

}
