% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-source_file.R
\name{html_tags}
\alias{html_tags}
\alias{default_tags}
\alias{basic_tags}
\title{HTML templates}
\usage{
default_tags(local = TRUE)

basic_tags()
}
\arguments{
\item{local}{TRUE / FALSE. If TRUE, the R base module is loaded from the
local file stored in the package, otherwise, the module is served via a
content delivery network (CDN).}
}
\description{
A list of 'shiny.tag' objects describing a HTML template. The
list must have signature / structure of a named list:
    \code{[head = [shiny.tag], body = [shiny.tag]]}
}
\examples{
str(default_tags())

str(basic_tags())

}
