% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_sketch_in_rmd.R
\name{insert_sketch}
\alias{insert_sketch}
\title{Insert a sketch into an RMD document}
\usage{
insert_sketch(id = "new_sketch", sketch_path)
}
\arguments{
\item{id}{A character string; the id for the <div> that contains
the sketch. This must match with the id used in the sketch file.}

\item{sketch_path}{A character string; the path to the sketch file.}
}
\description{
Insert a sketch into an RMD document
}
