% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_sketch_in_RMD.R
\name{insert_sketch}
\alias{insert_sketch}
\title{Insert a sketch into an RMD document}
\usage{
insert_sketch(file, id, ...)
}
\arguments{
\item{file}{A character string; the path to the sketch file.}

\item{id}{A character string; an identifier for the sketch file.}

\item{...}{(Optional) Other attributes to pass to iframes.}
}
\description{
Insert a sketch into an RMD document
}
\note{
This function creates temporary files at the working directory.
}
