% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-deparse.R
\name{list-of-deparsers}
\alias{list-of-deparsers}
\alias{dp_r_support}
\alias{dp_auto}
\alias{dp_dom}
\alias{dp_d3}
\alias{dp_macro}
\title{Low-level lists of deparsers}
\usage{
dp_r_support()

dp_auto()

dp_dom()

dp_d3()

dp_macro()
}
\description{
Support of R functionalities
}
