% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse.R
\name{make_deparser}
\alias{make_deparser}
\title{A constructor for a "typed" deparser}
\usage{
make_deparser(predicate_fun, deparse_fun)
}
\arguments{
\item{predicate_fun}{A function that takes a "lang" object and return a logical.}

\item{deparse_fun}{A function that takes a "lang" object and return a character string.}
}
\value{
A list; a deparser ready to be dispatched by "type".
}
\description{
A constructor for a "typed" deparser
}
