% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-components.R
\name{is_Family}
\alias{is_Family}
\alias{is_sym}
\alias{is_infix}
\alias{is_wrap}
\alias{is_call_for}
\alias{is_call_if}
\alias{is_call_while}
\alias{is_call_function}
\alias{is_call_return}
\alias{is_call_assignment}
\alias{is_call_assignment_auto}
\alias{is_call_break}
\alias{is_call_try}
\alias{is_call_tryCatch}
\alias{is_call_throw}
\alias{is_call_list}
\alias{is_call_df}
\alias{is_call_df_summarise}
\alias{is_call_df_mutate}
\alias{is_call_R6Class}
\alias{is_call_new}
\alias{is_call_typeof}
\alias{is_call_export}
\alias{is_call_let}
\alias{is_call_const}
\alias{is_call_var}
\alias{is_call_dataURI}
\alias{is_call_ifelse}
\alias{is_call_lambda}
\alias{is_call_pipe}
\alias{is_call_raw_string}
\alias{is_call_add}
\alias{is_call_subtract}
\alias{is_call_extract}
\alias{is_call_extractAssign}
\alias{is_call_extract2}
\alias{is_call_extract2Assign}
\alias{is_html_tags}
\title{Predicate for symbols, i.e. symbols or syntactic literals}
\usage{
is_sym(ast)

is_infix(ast)

is_wrap(ast)

is_call_for(ast)

is_call_if(ast)

is_call_while(ast)

is_call_function(ast)

is_call_return(ast)

is_call_assignment(ast)

is_call_assignment_auto(ast)

is_call_break(ast)

is_call_try(ast)

is_call_tryCatch(ast)

is_call_throw(ast)

is_call_list(ast)

is_call_df(ast)

is_call_df_summarise(ast)

is_call_df_mutate(ast)

is_call_R6Class(ast)

is_call_new(ast)

is_call_typeof(ast)

is_call_export(ast)

is_call_let(ast)

is_call_const(ast)

is_call_var(ast)

is_call_dataURI(ast)

is_call_ifelse(ast)

is_call_lambda(ast)

is_call_pipe(ast)

is_call_raw_string(ast)

is_call_add(ast)

is_call_subtract(ast)

is_call_extract(ast)

is_call_extractAssign(ast)

is_call_extract2(ast)

is_call_extract2Assign(ast)

is_html_tags(ast)
}
\arguments{
\item{ast}{A language object.}
}
\description{
Predicate for symbols, i.e. symbols or syntactic literals

Predicate for infix operators

Predicate for brackets

Predicate for the 'for' keyword

Predicate for the 'if' keyword

Predicate for the 'while' keyword

Predicate for the "function" keyword

Predicate for return

Predicate for assignments

Predicate for assignments

Predicate for the "break" keyword

Predicate for the "try" keyword

Predicate for the "tryCatch" keyword

Predicate for the "throw" keyword

Predicate for the "list" operator

Predicate for the "data.frame" operators

Predicate for the "summarise" operators

Predicate for the "mutate" operators

Predicate for the "R6Class" function

Predicate for the "new" operator

Predicate for the "typeof" operator

Predicate for the "export" operator

Predicate for the "let" operator

Predicate for the "const" operator

Predicate for the "var" operator

Predicate for the "dataURI" operator

Predicate for the "ifelse" operator

Predicate for the "lambda" operator

Predicate for the "pipe" operator

Predicate for the raw string operator

Predicate for the "add" operator

Predicate for the "subtract" operator

Predicate for the "extract" operator

Predicate for the "extractAssign" operator

Predicate for the "extract2" operator

Predicate for the "extract2Assign" operator

Predicate for the HTML tags
}
