% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rules.R
\name{r-to-js-rules}
\alias{r-to-js-rules}
\alias{basic_rules}
\alias{default_rules}
\title{Mapping R operators into JavaScript operators}
\usage{
basic_rules()

default_rules()
}
\description{
Mapping R operators into JavaScript operators
}
\note{
These functions are used as inputs to \link{compile_r} and \link{compile_exprs}.
}
\examples{
basic_rules()

default_rules()

}
\references{
R operators: \url{https://cran.r-project.org/doc/manuals/r-release/R-lang.html#Operators}

R infix and prefix operators: \url{https://cran.r-project.org/doc/manuals/r-release/R-lang.html#Infix-and-prefix-operators}

JavaScript operators: \url{https://www.w3schools.com/js/js_operators.asp}
}
