% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch.R
\docType{package}
\name{sketch-package}
\alias{sketch-package}
\alias{sketch}
\title{Interactive visualisation}
\description{
Creates interactive visualisation embeddable in R Markdown documents.
It implements an R-to-JavaScript transpiler and enables users to write JavaScript
applications using the syntax of R.
}
