% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-source_file.R
\name{source_active}
\alias{source_active}
\title{Source active file in 'RStudio'}
\usage{
source_active(...)
}
\arguments{
\item{...}{Optional arguments to pass to \code{source_r}.}
}
\description{
Source active file in 'RStudio'
}
\examples{
\dontrun{
# At 'RStudio', opens a 'sketch' R file in the editor, then
# run the following:
source_active()  # This launches the default HTML viewer.
}

}
